# shallow: stop checking output of base numeric fns