--- a/typed-racket-test/unit-tests/shallow-rewrite-expansion/base-env-numeric.rkt
+++ b/typed-racket-test/unit-tests/shallow-rewrite-expansion/base-env-numeric.rkt
@@ -1,0 +1,7 @@
+#lang typed/racket/shallow
+
+(let ((x : Nonnegative-Integer 5))
+  (= x 0)
+  (* x 1)
+  (+ x 1))
+