add the default "trust fn result" flag to the definitions at the top of base-env-numeric

thanks @shawnw for the report:
<https://www.reddit.com/r/Racket/comments/11dqjbb/experimenting_with_the_new_typed_racket_modes/?utm_source=share&utm_medium=web2x&context=3>