@samth 2023-03-05
Does this change any of the benchmarks you've done for shallow?