@shawnw 2023-03-05
> Does this change any of the benchmarks you've done for shallow?

I don't know; I just use OS packages for racket, never tried building from source to try an in-development version.