@bennn 2023-03-09
Yes! This makes a huge difference for `morsecode` (many seconds) and +1 seconds on a few others. Really should have caught this earlier. It's time to add Shallow to the perf-monitoring package and run that regularly (`gtp-checkup`).

Here are quick timings (ms) on the all-Shallow configurations for some of the GTP Benchmarks. Times are the average of 8 runs on a Cloudlab node (Utah m510).

| benchmark  | before ms (stddev)| after            |
| ---------- | ----------------- | ---------------- |
| morsecode  | 27925.00 (730.72) | 3316.38 (42.28)  |
| suffixtree | 10083.75 (360.02) | 8652.38 (132.21) |
| fsm        | 2879.00 (108.52)  | 1659.75 (42.32)  |
| tetris     | 3464.50 (102.95)  | 2372.00 (27.95)  |
| dungeon    | 903.13 (23.07)    | 876.13 (18.72)   |
| gregor     | 576.13 (11.00)    | 425.13 (22.41)   |
| synth      | 883.13 (13.93)    | 779.00 (25.36)   |
| snake      | 3794.13 (172.48)  | 3576.25 (45.23)  |
| kcfa       | 1385.00 (40.30)   | 1371.13 (20.55)  |
| lnm        | 677.88 (4.57)     | 649.00 (4.30)    |
| zombie     | 1365.00 (61.64)   | 1314.13 (19.22)  |
| mbta       | 2046.25 (39.84)   | 2013.63 (56.99)  |
| quadT      | 1469.13 (32.25)   | 1460.00 (19.32)  |
| quadU      | 1911.88 (9.53)    | 1944.13 (22.28)  |
| sieve      | 4689.75 (89.12)   | 4048.00 (72.52)  |